requirements.txt
Flask==3.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# 开发依赖
# pytest==7.4.0
# black==23.7.0
# flake8==6.0.0